requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
asn1crypto==1.5.1
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
contourpy==1.1.1
cryptography==45.0.5
cycler==0.12.1
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.3.0
filelock==3.16.1
fitz==0.0.1.dev2
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
groq==0.29.0
h11==0.16.0
holidays==0.58
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib-resources==6.4.5
isodate==0.6.1
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
looseversion==1.3.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
narwhals==1.42.1
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
prophet==1.1.7
protobuf==5.29.5
prov==2.0.1
pyarrow==17.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
pydeck==0.9.1
pydot==4.0.1
pygments==2.19.2
PyJWT==2.9.0
pyopenssl==25.1.0
pyparsing==3.1.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
referencing==0.35.1
requests==2.32.4
rich==13.9.4
rpds-py==0.20.1
scipy==1.10.1
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.14.0
sortedcontainers==2.4.0
stanio==0.5.1
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.4.2
tqdm==4.67.1
traits==6.3.2
typing-extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
watchdog==4.0.2
zipp==3.20.2
